perf(bundle): move zod off boot path into lazy ipc.validate chunk
zod was statically imported in api/ipc.ts, which is transitively required
by bindings/index.ts → Shell at startup. Moves the schema + helpers to
ipc.validate.ts and pre-warms it with a fire-and-forget dynamic import so
the validator is available before any IPC error can arrive. On a cache miss
unwrap skips validation and rethrows unchanged.

ipc.test.ts updated to import ContractErrorSchema/validateContractError
from the new module directly.
