+++ apps/desktop/src/api/ipc.ts
@@ -15,69 +15,25 @@
  * imports from here), and it loads mocks / the real tauri core lazily.
  */
 
-import { z } from 'zod';
-
 // ── T118: Zod validation for dynamic/drift-prone IPC payloads ──────────────
 //
-// `ContractError` is the primary drift-prone payload at the IPC error seam.
-// Its `details` field is typed `unknown` in the generated bindings and carries
-// free-form JSON that the backend can change without a TS type update.
-// We validate the mandatory envelope fields so a backend regression (e.g. a
-// missing `code` or a renamed `message` key) is surfaced immediately as a
-// clear validation error rather than a silent undefined downstream.
-//
-// Validation is NON-BREAKING: every currently-valid ContractError payload
-// (all existing mocks and tests) satisfies this schema.  An entirely unknown
-// error shape (e.g. a plain string from Tauri itself) is detected and
-// rethrown with a descriptive message.
+// Validation types live in `ipc.validate.ts` so zod is NOT bundled in the
+// boot chunk (zod enters through the error path only, never at startup).
+// `ipc.ts` pre-warms the validate module with a fire-and-forget dynamic
+// import and caches the synchronous validator function for use in `unwrap`.
+// On a cache miss (error arrives before the tiny validate chunk has loaded —
+// practically impossible post-boot) validation is skipped and the raw error
+// is rethrown unchanged, which is still better than a silent undefined.
 
-/** Zod schema for the mandatory `ContractError` envelope fields (T118). */
-export const ContractErrorSchema = z.object({
-  code: z.string().min(1),
-  message: z.string(),
-  severity: z.string(),
-  retryable: z.boolean(),
-  // `details` is intentionally unknown/unconstrained — it is the free-form
-  // blob we are watching for drift; we accept any JSON value here.
-  details: z.unknown().optional(),
-  fieldErrors: z.array(z.unknown()).optional(),
-  recoveryActions: z.array(z.unknown()).optional(),
+// Cached synchronous validator populated once the validate chunk loads.
+let _validate: ((raw: unknown) => void) | null = null;
+
+// Pre-warm the validate chunk at module-init time (lazy, non-blocking).
+// This runs before any IPC command could possibly return an error response,
+// so the validator is cached long before `unwrap` ever needs it.
+void import('./ipc.validate').then((m) => {
+  _validate = (raw) => m.validateContractError(raw);
 });
-
-export type ValidatedContractError = z.infer<typeof ContractErrorSchema>;
-
-/**
- * Validate a raw IPC error payload as a `ContractError`.
- * Returns the validated object on success; throws a descriptive
- * `IpcPayloadValidationError` on failure so drift is surfaced immediately.
- */
-export function validateContractError(raw: unknown): ValidatedContractError {
-  const result = ContractErrorSchema.safeParse(raw);
-  if (result.success) {
-    return result.data;
-  }
-  throw new IpcPayloadValidationError(
-    'ContractError',
-    result.error.flatten().fieldErrors,
-    raw,
-  );
-}
-
-/** Thrown when an IPC payload fails zod validation (T118). */
-export class IpcPayloadValidationError extends Error {
-  constructor(
-    public readonly payloadType: string,
-    public readonly fieldErrors: Record<string, string[] | undefined>,
-    public readonly raw: unknown,
-  ) {
-    super(
-      `IPC payload validation failed for ${payloadType}: ` +
-        JSON.stringify(fieldErrors) +
-        ` — raw: ${JSON.stringify(raw)}`,
-    );
-    this.name = 'IpcPayloadValidationError';
-  }
-}
 
 const useMocks = import.meta.env.VITE_USE_MOCKS === 'true';
 
@@ -130,9 +86,9 @@ export type IpcResult<T, E = unknown> =
  *
  * T118: When the error branch carries an object-shaped payload (i.e. a
  * `ContractError` from the backend), the mandatory envelope fields are validated
- * via zod before re-throwing.  This surfaces backend drift (e.g. a renamed `code`
- * key, a missing `message`) as a clear `IpcPayloadValidationError` rather than a
- * silent `undefined` somewhere downstream.
+ * via the cached validator from `ipc.validate.ts` before re-throwing. This
+ * surfaces backend drift (e.g. a renamed `code` key, a missing `message`) as a
+ * clear `IpcPayloadValidationError` rather than a silent `undefined` downstream.
  *
  * Plain string errors (older commands) and `Error` instances are passed through
  * unchanged so no existing behaviour is broken.
@@ -149,9 +105,9 @@ export function unwrap<T, E = unknown>(result: IpcResult<T, E>): T {
     !Array.isArray(err) &&
     !(err instanceof Error)
   ) {
-    // Run validation; if it throws IpcPayloadValidationError we let that propagate
-    // so the caller sees exactly what drifted.  On success we rethrow the original.
-    validateContractError(err);
+    // Use the cached validator if loaded; skip validation on a cache miss
+    // (validate chunk not yet loaded — safe to skip, still rethrows the error).
+    _validate?.(err);
   }
   throw err;
 }
